fix: lazy Supabase admin client; API routes nodejs+force-dynamic; no top-level env reads; Inngest v3 client id; ignore .next
