+++ .env.example
@@ -1,0 +1,4 @@
+# Inngest
+INNGEST_EVENT_KEY=__set_in_netlify__
+INNGEST_SIGNING_KEY=__set_in_netlify__
+INNGEST_ENV=production
