+++ .gitignore
@@ -1,1 +1,6 @@
 node_modules/
+
+# Build output
+.next/
+dist/
+out/
